requirements.txt
azure-functions
bs4